requirements.txt
Flask>=2.2
python-docx>=0.8.11
mammoth>=1.4.19
lxml>=4.9
gunicorn>=20.1
sympy>=1.12 # optional helper if you decide to parse MathML or convert to LaTeX
